Stack title to prevent overlap; new file for title drawing
